+++ FootballBet.Repository/Repositories/Interfaces/IUserRepository.cs
@@ -5,5 +5,5 @@ namespace FootballBet.Repository.Repositories.Interfaces;
 public interface IUserRepository
 {
     public Task<ApplicationUser?> GetApplicationUserById(string userId, CancellationToken ct);
-    public Task<double> UpdateBalance(string userId, double balance, CancellationToken ct);
+    public Task<decimal> UpdateBalance(string userId, decimal balance, CancellationToken ct);
 }
